+++ tests/cli/commands/test_webserver_command.py
@@ -226,6 +226,7 @@ class TestCLIGetNumReadyWorkersRunning:
             assert self.monitor._get_num_ready_workers_running() == 0
 
 
+@pytest.mark.db_test
 class TestCliWebServer(_ComonCLIGunicornTestClass):
     main_process_regexp = r"airflow webserver"