+++ airflow/providers/mongo/CHANGELOG.rst
@@ -38,6 +38,7 @@ Features
 .. Below changes are excluded from the changelog. Move them to
    appropriate section above if needed. Do not delete the lines(!):
    * ``Pre-upgrade 'ruff==0.0.292' changes in providers (#35053)``
+   * ``Prepare docs 3rd wave of Providers October 2023 (#35187)``
 
 3.3.0
 .....